Respond only to challenges for packets sent
Each outgoing packet is hashed and we store the hash in a ring buffer.
If the sender receives a challenge from a router, it checks the ring
buffer for a matching packet to ensure that the challenge is for a
packet sent by it. If it is, a response to the challenge is generated
and sent to the challenger.