Adding check for timestamp in migrateack
When a migrate message is sent for a migration, we record the timestamp
and it serves as a nonce as well as to protect against back to back
migrations. On receiving a migrate ACK, we check the timestamp returned
in that message against the one from the most recent migration. If it
matches, we should accept the migration as complete. Currently, we just
warn the user if they don't match. One reason for them to not match is
if the mobile host hops to another network before the first migrateack
is received, prompting a new migrate message to be sent.