migrateack_waiting was uninitialized
In DAGinfo created for a socket, migrateack_waiting was undefined
causing it to have undefined behavior later in the code. Simply
initialize it to false when the socket is first created. It is not
relevant for anything other than stream sockets, but should not
negatively affect them if it is initialized to false irrespective of
whether it is used or not.