For ChallengeResponder ether hdr not stripped
All outgoing packets are hashed so a challenge can be responded to later
for any of them. The packets being hashed included the ethernet header
(14 bytes) but the ChallengeSource element on the other end received it
stripped of those 14 bytes. This change strips the first 14 bytes for
the ChallengeResponder just like they are for ChallengeSource.