select based retries for nameserver registration
When we register a name or service with the nameserver, we were sending
one UDP based registration message and then waiting for the nameserver
response. Problem is, the registration message can be dropped by the
network or by the intrinsic security implementation. In either case, a
response from the nameserver never comes causing code to wait
indefinitely. This new implementation based on select provides for a
fixed number of retries with timeouts that will cause the code to return
with an error message in about a minute if there is no response from the
nameserver.