Keys and conf files persist multiple invocations
We used to autogenerate the config files on each invocation from the
templates. This does not play well with intrinsic security that ends up
generating a new set of keys and assigning new XIDs every time the
network is restarted. This change addresses the issue by simply
disabling the template engine when the config files already exist.