clean keyword deletes all config and key files
Removes all configuration done by previous invocations of bin/xianet.

It also removes the resolv.conf file but we have to rethink that. In
case we are running on multiple machines, the resolv.conf has to be
removed only when clean is called on the machine running the nameserver.
For now we just remove it and the user is expected to copy it back if it
was supposed to stay unchanged.