Xsend on raw socket forwards raw packet now
When a raw packet is delivered to user space and then then returned back
via Xsend, we just forward the packet as is, without modifying the
header or contents. The application holding the raw socket is
responsible for providing a routable raw packet to Xsend.
We cannot allow Xsendto because a destination is needed and the
application may not be able to easily provide a valid one. If we do want
to use Xsendto, the destination options can probably be ignored and just
a check for raw socket may allow us to forward packet as is.