Rendezvous daemon modifies and forwards correctly.
The rendezvous service data plane finds the AD and HID in the packet.
Then it finds the HID in its internal table and the corresponding new
AD. The new AD is substituted for the old one in the header. The last
pointer is set to -1 (the source node) and the packet sent back out on
the network.