Update peer address on receiving SYNACK
Since the rendezvous service may change the destination DAG for a packet
mid-flight, if a SYN packet reaches a server, the server sets its
address in the outgoing SYNACK to reflect the DAG that got the packet to
the server in the first place. This SYNACK reaches the client. At this
point the client must set its peer DAG to that of the src DAG of the
SYNACK packet. This change makes that last step happen.

So at this point, the first SYN packet for a mobile server goes through
the rendezvous service, but all subsequent packets reach the server
directly.