Disable removal of CID packet from cache
For some reason, we were removing a valid CID packet from cache after it
was delivered to the application. However, the corresponding map that
held the status of the packet was not updated. Causing click transport
XreadChunk() code to believe that the CID is available when it was
requested for a second time during the same file transfer in xftp.
This change could possibly cause a memory leak. If it does we will have
to find the underlying logic for why the CID was being dropped from
memory and how the corresponding map should be updated to reflect that.