SYNACK verifies DAG changed by rendezvous service
When a SYN request is received after modification by the rendezvous
service, the SYNACK sent to the client includes a verification payload
signed by the mobile service and the mobile service public key. The
client checks to make sure that the SYNACK came from the same DAG it
sent the SYN to. If not, it looks for a signed verification from the
mobile service before accepting the rendezvous modified DAG. Only after
the signature is verified and the public key matches the remote SID do
we allow alteration of the mobile service DAG at the client side.