Use new xs_pubkeyMatchesXID() for security messages
The new xs_pubkeyMatchesXID() function simplifies code to verify that
the public key and the corresponding HID/SID match. It verifies that the
hash of public key argument is the same as the hash portion of the XID
string. This change makes the rendezvous control message handler and the
XIAChallengeSource verifier use the new function.