Harden signing code
If the user program deletes the key files corresponding to a socket that
is still active, the server may still be accepting connections on it. If
a connection is accepted and it was coming through the Rendezvous
service, we return a signed SYNACK so the remote end can accept a DAG
different from the one it used to initiate the connection. This signing
will fail because the key file is missing and it was causing Click to
crash. Now we simply return from the function. It may be appropriate to
return an error code to the user space in such a case but that is not
done by this change.