Return error to Xaccept on RV DAG signing issues
The API Xaccept sends a READYTOACCEPT request to Xtransport. The
Xtransport::XreadyToAccept notifies the API Xaccept when a SYN is
received for this connection and it then sends an XACCEPT request to
Xtransport::Xaccept to handle the incoming packet. If this incoming
packet is coming through the RV service, the SYNACK to be sent in
response is signed by the private key of the accepting SID and its
public key is also included in the SYNACK. If we run into any issues,
such as the key files being missing or the system going out of memory,
we should notify the application that the Xaccept request has failed. It
is up to the application to call Xaccept again in a loop if it wants to
continue accepting connections after this error.