Moving towards XHCP client with multiple interfaces
The client currently just prints the beacon contents but is able to get
the beacons from multiple interfaces. It does not act on the beacons so
is completely broken in its current state. We need to build Click
support for the actions to work and that is next on the list of TODOs.