xtransport now accepts network_dag and hid
It forms the local address internally every time the network_dag in
configured via a write handler. It requires the HID to be configured
before an attempt is made to change the network address.

We envision the HIDs to be assigned at setup time and network_dag to
change as the XHCP beacons arrive from a new network when the device
appears on the network. This code will have to change if we want HIDs to
be configured anytime after setup. Just need to make sure that the local
address is correctly maintained.