No more xia_address.click in read_topology
We no longer create xia_address.click in read_topology.
The processing loop is now cleaner and writes nodes.conf and
sockconf.ini in a single pass.
We also don't create a resolv.conf for local topologies. It will be
created after Click is started and its elements have ADs and HIDs
assigned.