Config files from templates now go to etc/click
The host/router.click files generated from templates were going to
etc/click/templates. Once a template is used to generate a click file,
the click file is not a template, so it shouldn't be put in templates
directory.
Need to change bin/xianet to find the file in correct location next.