Nameserver will run on router0 for local topology
We will no longer be able to configure a topology with the nameserver
running on a host. This is because of the multihoming changes.
When we process the topology file, we do not hae the AD assigned for the
network on which a host running the nameserver would be running. Finding
that information would involve parsing the topology and we just chose to
save development time by not doing that right now.