New script assign_addresses now assigns HIDs
bin/xianet calls bin/assign_addresses to assign HIDs to all elements
that need it. It will also assign ADs for routers in the near future.
Also included in this change is the ClickControl library, that allows
communication with Click. It will provide functionality to change
network DAG at run-time in the future.