XARPQuerier was messing with the configure input
Not quite sure what that code was trying to do but it was pushing an
entry back onto the conf Vector passed to configure making it too many
arguments when it went to parse the mac address. The removal of HID
argument seems to have triggered this parsing issue. Need to talk with
Dan about this.