Xtransport uses HID DAG until network DAG is known
As a chicken and egg problem, we need an application(XHCP client) to
send out packets to the gateway router and name server, but the entire
DAG is not known until after we start receiving beacons. To solve this
and other such problems we will use the *->HID DAG as the source address
on these outgoing packets. That allows the router on the other end to
still route packets back to our application.