Remove reverse mapping from DAG to interface
The DAG is the same for all interfaces on a router which is contrary to
the original plan where we said the DAGs for all interfaces will be
unique. This prevents us from having a simple DAG -> interface mapping
table, but since it is not used anywhere, I just commented out the code
to handle that table.