XHCP Server sends router's DAG in beacons
All other functions were updated to receive a DAG for the router and
replace the HID in the DAG with the HID for the local interface to form
a complete DAG for a host. This is necessary because passing a
network_dag with multiple paths reaching the destination HID was not
possible. This change is large and includes:
* XHCP server sends router's DAG in a serialized string
* XHCP client receives router's DAG as a string
* XHCP client calls new XchangeAD to update the DAG
* XHCP client keeps router's DAG in its state, instead of just AD
* XchangeAD API was changed to accept the new dag and interface number
* XupdateAD in Click was updated to accept the router's dag and use it
* The protobufs to exchange between API and click were updated
* XreadLocalHostAddr now returns the DAG, not its individual components
* All apps that use XreadLocalHostAddr had to be updated

Note: New XupdateAD in click is not doing active session migration yet
Note: Since XHCP client keeps router's DAG, if the host moves between
	two routers on the same AD, it will trigger a migration event
	even though the resulting DAG for the host may be the same.
	This, because only the RHID will change in DAG, which is then
	replaced by the host to form its DAG.