Enable migrate messages based on interface
The current logic is to determine the interface for all new stream
connections based on the DAG being bound to, accepted on, or being
connected from. If we are able to figure out which interface the DAG
corresponds to, we note that in the per socket state in Click. On
migration, we migrate the connections corresponding to the interface
that will be using a new DAG.