Xgetaddrinfo now returns a list of addrinfo structs
With multihoming support, Xgetaddrinfo now returns a list of addrinfo
structs based on all available interface DAGs if the user has not
provided a "name" argument to be used as system DAG.
This can cause problems for applications that assume only one available
interface. We may have to return the "default" interface DAG as the
first entry in the list to protect these applications.