RVD won't Xgetaddrinfo() for resolv.conf DAGs
We used to read the Data-plane and Control-plane DAGs from resolv.conf
and extract the SIDs from them. Then later, RVD would call
Xgetaddrinfo() with the SIDs as arguments to form the Data-plane and
Control-plane DAGs instead of using the ones in resolv.conf.
Now, if the user provides SIDs as command-line args, we build DAGs from
those. Otherwise, we look for DAGs in resolv.conf and use those. Lastly,
if no SIDs were provided and no resolv.conf entries exist, we build
DAGs based on randomly picked SIDs. That last option is only for testing
purposes.