Include RV Control DAG in XIAInterfaceTable
Each interface must keep track of last known rendezvous service so it
can be notified of latest addresses, even from networks that don't
provide rendezvous services of their own.

In future, we may want the rendezvous service to be configurable. With
this current design, the client must be started in their home network
providing rendezvous service or visit another network providing that
service. It is not clear if a foreign network would like to keep
providing rendezvous service to clients that are no longer in their
control. So the best may be to allow the user to configure one at
bootstrap time with a script and update as necessary to a foreign one if
they are too far away from home ... or whatever the policy is.