XupdateRV now correctly updates RV control plane
With the multi-homing changes XupdateRV must now pass in the interface
number as well as the rendezvous control plane DAG it received via the
XHCP beacon in XHCP client. This Rendezvous DAG is stored in
XIAInterfaceTable. If the host moves to a new network without RV
service, the old rendezvous service will be notified of the change
through its control plane.