XIANetJoin element now in place
A new XIANetJoin element. Currently plugged into the XIALineCard. This
is not the final location; most likely will be split into two elements.
One piece will sit in the LineCard to create Ethernet Frames going out
on wire and a second central piece to sit in Xtransport to
multiplex/demultiplex the LineCard elements.
The current implementation is only intended to be a proof-of-concept.