Adding hostname argument to xnetjd script
The startup script bin/xianet expects each running service to have the
hostname as one of its arguments, so it can control the service and kill
it when needed. This change adds a --hostname argument to xnetjd even
though it is not actually used.
A later task is to allow more than one xnetjd instances to run on the
same machine for local topology etc.