NetjoinBeacon calculates its own ID now
We used to hash a serialized ndap_pb2.NetDescriptor object to create an
ID for the beacon. This no longer works because the NetDescriptor
contains a nonce that changes in every beacon.
The NetjoinBeacon.get_ID() method copies the NetDescriptor into a
separate object and then clears out the nonce before taking a hash on
it. Inefficient, yes. One solution is to change NetDescriptor to
contain all non-nonce data in a separate submessage. Future work...