Gave the SCION forwarder access to the HID route table.
At this point in the code, the scion header has port numbers
to get to the next node is the route. We can't forward without
knowing the nexthop XID, so we find the port in the HID table
and return the associated nexthop.
When the SCION header changes to IDs or something else, these changes
may need to be replaced.