Refactored to use Authsession
In addition to using NetjoinAuthsession for all crypto needs, other
changes include:
The beacon from Announcer contains a NetDescriptor, calling it beacon
was confusing because we use that term for the wrapper NetjoinBeacon.
The announcer creates an Authsession and a reference to it is passed to
the receiver. Which in turn will use the announcer's Authsession when it
receives a response to a beacon sent by the announcer.
NetjoinBeacon was doing a bunch of key management. That has now moved to
NetjoinAuthsession. Only the raw public (verify) key is passed to
beacon object for inclusion in the announcement.