Filter out non-XIP pkts to XIAChallengeResponder
Outgoing packets are funneled to XIAChallengeResponder. Which in turn
keeps a hash of the recently seen packets and uses this value to respond
to any AIP style challenge from a router on the path. The Netjoin
packets were also getting forwarded to XIAChallengeResponder, which
obviously didn't know how to handle these raw packets. Simply added a
filter to only pass XIP packets. This may bite us if the challenge
response protocol was also responding to XARP packets. We'll find that
out and fix it if necessary.