AP NetjoinSession gets a copy of announcer's auth
The auth session at the access point side is owned by the announcer
object. When we create a joining session on receiving a handshake one
from a client, we provide a copy of the auth session to this newly
created session. This allows the session to replace the client's public
key without affecting sessions with other clients.

The advantage/disadvantage of this solution is that we are creating
copies of the Announcer's AuthSession private key also. It is good
because Announcer can freely change its ephemeral private key without
affecting existing sessions. It is bad because we want to minimize
copies of private keys to minimize leaking them.

There are of course other solutions:
1. Break NetjoinAuthsession into a local and remote part.
2. Atomically update their verify key and use the auth session.
3. Atomically provide their verify key to auth and get a Box for encryption

We can revisit this decision later if the current solution is found
insufficient for our needs.