NetjoinXIAConf reads XIA conf & provides utilities
Read in the XIA config files from the local environment to extract info
like HID, Public key corresponding to the HID.
It also provides utility functions to convert HID and pubkey to binary
for going over the wire and back to hex at the destination.
Soon, I'll add signing capability, so we can use the HID private key to
sign a challange from the access point.