Add routing table updates to Xupdatedag
Xupdatedag was updating a bunch of click elements with the new DAG but
was not updating the routing tables. This task was delegated to the XHCP
Client. However, there is no reason to do the routing table updates in
XHCP or as part of network joining in user space. So I consolidated
everything in one place, Xupdatedag().