Xgetaddrinfo was not returning looked up dag
Xgetaddrinfo was doing a nameserver lookup, but was not including the
looked up dag in the returned addrinfo structs if the service name was
not provided by the caller.

In case of a name lookup by human-readable string, the service type is
not provided, resulting in a null lookup result.