Fix nonce usage.
The nonce is already included in each encrypted message by the crypto
library. We also include it separately as part of the handshake
messages, which we can avoid to save bytes.

Due to a misunderstanding, I had the nonce also acting as a challenge
for the next message. However, that is not the correct usage for the
nonce used for encryption of each message. So this change fixes the
logic to stop using the same nonce as a challenge for the next outgoing
message.