Generate keys using openssl executable
We have had compatibility issues with python-crypto and python-rsa
generated public keys being exported. One exports in x.509 and the other
exports in pkcs#1. We were using python-crypto but it is a pain to cross
compile for mips. Since we already built openssl and use openssl for all
our C code, it makes sense to just use that instead of python libraries
to generate the keys in the first place.