Simplified Arada OBU/RSU deployment steps.
This scripts now assumes that the arada-install.sh script has been
successfully run and it has prepared the necessary packages for going on
to the Arada hardware. We were not saving much with preparing a tarball
and unpacking it, so I decided to switch to a simple recursive scp
instead. That also ensures that the user is asked for the password once
instead of twice.