The -s option to hostname doesn't work on Arada
hostname -s expects to read an entry for the hostname from /etc/hosts
but Arada doesn't have an entry there and adding a persistent entry in
the temporarily mounted filesystem (tmpfs) is not a decent option.
So for now, we just have a hack that calls "hostname" without -s if the
machine we are running on is a mips machine.