Include daemons/xnetjd in python path
On x86 boxes, we symlink bin/xnetjd to daemons/xnetjd/xnetjd.py. So when
we run ./bin/xnetjd, it is able to find all the modules correctly in the
daemons/xnetjd/ directory.
On Arada OBU, the USB drive is VFAT formatted. So no symlink support.
Then running ./bin/xnetjd results in not being able to find necessary
modules. This script change detects the correct path where we expect to
find the modules and add it to the sys.path used to find them.