Disabling Crypto use for AIP source verification
Since we don't have Crypto module cross-compiled for the mips based
Arada boxes, we are commenting out the AIP source verification piece of
the Network Joining protocol. This change needs to be reverted or
replaced by the use of a different cryptography module.
Ideally, we should just be using libsodium/PyNaCl for all crypto. That
is what we use for encryption/signing of network joining messages.