getAradaInterfaces method to deal with DSRC ifaces
On Arada boxes our clever xconfig.py script was failing to figure out
which of the several available interfaces to use for the DSRC
functionality. This change puts in the following logic:
Only look at wifi0 and wifi1 as the potential interfaces to use.
Read in and convert mac address of each interface into an int.
The interface with the lower integer representation of mac address is
the one we use.
This is really a hack until we have a better way of choosing the
interface that is going to use the DSRC channel we are setting up.