Reduce Ethernet retransmit rate
For now, the netjoin code is just using the ethernet l2 handler for DSRC
interfaces. This handler was retransmitting at 0.1 sec and that is way
too fast for the Arada boxes running our stack. So this change reduces
the rate to 1 second retransmissions.
Soon, we need to send out different beacons on different interfaces
depending on the l2 they support.