The DSRC template file for a host
On Arada boxes, we use a separate template file that configures only one
interface, the wifi interface corresponding to the DSRC device that
should be used for communication.

The xconfig.py script that generates the configuration from this
template is not smart enough (yet) to generate a multihoming config. It
should not be too hard to generate a multihoming config that enables the
device to use one of the ethernet interfaces, it's just that I have not
got around to do the necessary changes yet.