Ignore network announcement from peer routers
We wanted to allow a router to join an upstream network as a client.
However, this functionality can only be implemented if we explicitly
state which ports are downstream and which are upstream. Otherwise,
two peer routers see announcements from each other and try to join
their peer's network as a client and end up with routing tables and
DAGs in a bad state.