Temporary hack to work around xmit buffer overflow.
The click TCP code was discarding data if the xmit buffers were full.
The correct fix is to deal with this in click and not return until we've
dealt with the data. But for now, click will return 0 to Xsend to let
the app know it wasn't able to process the data.