Allow joining previously known network
The policy module used to keep track of known networks or beacons without
regard for the interface they arrived on. It also did not expire these
known interfaces. The result was that:
* Two interface could not join the same network
* If a network was joined before a migration, on return it would be ignored.
This change fixes both issues by making the policy module aware of the
interface on which a beacon arrived. So a per-interface joining decision
is made. If an interface connects to a new network, the old entry for that
interface is discarded, allowing the policy module to make a new decision
if we return to the old network in future.