Make logging work better.
Moved the xlog call to follow the speep instead of preceeding it.
Doing this makes logging to the console more consistant. It's possible
some startup messages sent before the sleep ends will hit syslog,
but not the console. This can be fixed by editing the verbose parmeter
in [host|router].click.