An "affinity" metric to choose among available networks
This change introduces an "affinity" number that is maintained per
interface per beacon. So if an interface receives multiple beacons
an affinity number is assigned to each based on reception frequency
and packet loss. The policy module can then allow switching to the
best of available networks. For now, we just use the frequency of
beacon reception in the affinity calculation. This can later be
extended to include RSSI, l2 characteristics etc.