option length now in dwords or multiples of 4 bytes
In preparation for the large migrate option (~700 bytes) we changed
the meaning of the option length field. It is now to be interpreted
as the number of dwords or multiples of 4 bytes. So a 4 byte option
would have a length of 1. The special options NOP and EOL are still
one byte options because they don't specify a length.
No extra bytes were added to the actual packets, we simply took out
the NOP options being used to pad for 4-byte alignment.