Put the router's rendezvous DAG in resolv.conf
We used to just store the rendezvous fallback path in resolv.conf
but need to store the entire path, so xnetjd can pass it as-is to any
clients that need it, without them having to assemble complex DAGs.
This may interfere with API calls like Xgetaddrinfo that used to stitch
the rendezvous path as a fallback.