DAG comparison now possible
As long as two DAGs are logically representing the same address they
will now be considered equal by the == operator.
We accomplish this by walking the DAG depth first from source to sink,
followed by all possible fallbacks from each node encountered.
This walk is recursive and there is currently a hard-limit to break
infinite recursion if we encounter it.