Remove policy, capacity and unregister_meta()
The xcache controller and xcache client were accepting new argument
capacity and creating an instance of a policy object. It also had an
unregister_meta() function. All these changes are going to be superceded
by work on the new cache policies and management by Dan and Heron. So
these changes were reverted to match the xia-v2 branch.