Porting sid_routing daemons to xia-v2.
Because the branches have diverged so much, it wasn't possible
to merge sid_routing with v2, so this code has simply been lifted
from that branch. Go there for history on this code.

Additionally, this code is being refactored from 2 route daemons
into a single one and the code has been heavily modified, again
losing the history. Currently it will run as either a controller,
or a router but not as both at the same time yet.

Still to come, dual mode operation, and legacy routing support.