Fix intent node removal in Graph
We return an error if the caller removes removal of last remaining node.
For each node, if edge to intent exists, erase and skip to next node.
We were not skipping to next node and iterator was invalid after erase.
Include intent node in search for intent xid node.
Also fixed an incorrect 'if' check for intent sid removal.