Disable purging hosts without heartbeat in xrouted
The host xhcp client used to send a periodic heartbeat to xrouted.
However, with new network joining, that task is not being done by
any entity. So for now we disable removal of host route from xrouted
due to lack of heartbeat. This will be fixed when the new routing code
to allow multiple routers in an AD becomes active.