XID type on wire has to be in network byte order
The routing code expects the XID type to be in network byte order.
The XID struct was always storing in network byte order. The Graph
Node class does not store in network byte order. We just convert it
to network byte order when writing to wire format for XIA headers and
for sockaddr_x.