Reorder statement order so cli/serv apps can run on the same node.
Because of how click works, it would go 70,000+ stack frames
deep when a client connected to a server on the same host.
This change reorders some statements during the tcp handshake
and appears to fix the problem. It's possible that in some edge
conditions there may be additional sequencing issues yet to be
uncovered.