Set host routes here instead of via the hack in click.
Host routes were being set by the routing table code when braodcast
messages were received. This didn't make much sense and won't work
anymore now that we are eliminating the broadcast HID. Main change here
is to use Xrecvmsg instead of XrecvFrom so that we can get the port
number along with the packet.