Don't need to set values in route tables anymore.
The default route engine doesn't handle broadcasts anymore so don't need
the HID of the machine to be set. Additionally, since the redirect logic
has been removed, we don't need to set the DAG either.

clickcontrol is still setting the HID in the FID engine, but this should
be temporary.