Provide headroom when creating packets in xstream
If a packet was created in xstream it will almost certainly get an
XIA header and a mac header. So provide plenty of room for those headers
to avoid expensive copying.

This hides a bug where the expensive copying causes data corruption.
If the headroom is set to 0, somehow, we end up with an occasional
packet containing 23 bytes of XIA header overwritten with some other
data. Between multiple corrupt packets only about 6 of those bytes change.