Only controller reads DAGs from file now
Routers will join upstream/peer routers and receive their DAGs,
nameserver DAG and Rendezvous DAGs from Join response (HS2).
Currently, we configure a router just like a host on receiving HS2 by
calling XupdateDAG, but we may need to limit that to doing it only if
the router is not already configured.
We are also setting up forwarding between joining routers just like we
do for the hosts but that may not be needed with FIDs in place.