Routers now announce network after joining it.
After joining a network, if we are running as a router, we announce the
newly joined network to other routers and hosts.

Since there is no check to see if we are already announcing the network,
we could end up starting a new announcer for each peer router that we
join. I'll look into adding a fix for that.