Fix xcache core dump.
 If a chunk fails, it's left open in the cache in case a stray
packet needs to be resent. But if a new request for the same chunk comes
in, a bug allowed this new session to try to reuse the chunk in waiting
state. SID tracking was supposed to prevent this, but was using the
chunk's SID instead of client's SID. Fixed this.

Also removed a bit of debug logging.