Enhance ns to store type of record.
The wrapper needs to be able to register an IP:port and host:port
identifier for the same DAG. But needs to do a reverse lookup and always
get one or the other back, and not rely on the ns choosing the right
one. Save the type of name along with the DAG so we can filter
correctly.

NOTE: this is a hack, and is another argument for revamping this code!