Fix depth_first_walk and compare_except_intent_AD
depth_first_walk() was dealing with two indices. Internal index for
nodes_ and user-facing dag string index. There was confusion with the
source index. It is sometimes 0, but sometimes the last node in nodes_.
So changed the function to only use the user facing dag string index
returned by get_node().

compare_except_intent_AD() intended to create a copy of the other graph
and modify it to match 'this'. However, the '=' operator doesn't do a
copy. It just creates a new object with reference to same underlying
data structures in the other object. Fixed by creating an explicit copy
by invoking the Graph::Graph(Graph &g) constructor.