XAI_DAGHOST must be set if DAG is provided
If Xgetaddrinfo is called with 'name' argument pointing to a DAG string
it must have a corresponding 'hints' argument with ai_flags having
XIA_DAGHOST flag set. This was acceptable but not mandatory but due to
some recent change it seems it is required.

Also changed buffer size for xia_ntop to allow for large RV DAG.