Removed obsolete XrendezvousUpdate() API
This API allowed an application to send control-plane updates to the
rendezvous service to notify it of the host's new address after
migration. However, this is now done from inside Click because
intrinsic security requires this update message to be signed by the HID
which is not accessible to user-space applications.